requirements.txt
imapclient==3.1.0
pathvalidate==3.3.1
pyyaml==6.0.3
smbprotocol==1.15.0
rich==14.2.0

# Testing
pytest==9.0.2
pytest-cov==7.0.0
